+++ src/scoped_timer.hh
@@ -2,16 +2,16 @@
 
 typedef struct scoped_timer
 {
-    scoped_timer(double& seconds) : seconds(seconds) {
+    scoped_timer(float& seconds) : seconds(seconds) {
         t0 = std::chrono::steady_clock::now();
     }
 
     ~scoped_timer() {
         auto t1 = std::chrono::steady_clock::now();
-        std::chrono::duration<double> diff = t1 - t0;
+        std::chrono::duration<float> diff = t1 - t0;
         seconds = diff.count();
     }
 
-    double& seconds;
+    float& seconds;
     std::chrono::steady_clock::time_point t0;
 } scoped_timer_t;
