+++ src/vectors.cc
@@ -27,7 +27,7 @@ vec3_t operator-(vec3_t a, vec3_t b)
     return a + (-b);
 }
 
-vec3_t operator*(vec3_t a, double b)
+vec3_t operator*(vec3_t a, float b)
 {
     a.x *= b;
     a.y *= b;
@@ -35,19 +35,19 @@ vec3_t operator*(vec3_t a, double b)
     return a;
 }
 
-vec3_t operator*(double a, vec3_t b)
+vec3_t operator*(float a, vec3_t b)
 {
     return b * a;
 }
 
-vec3_t operator/(double a, vec3_t b)
+vec3_t operator/(float a, vec3_t b)
 {
-    return b * (1.0 / a);
+    return b * (1.0f / a);
 }
 
-vec3_t operator/(vec3_t a, double b)
+vec3_t operator/(vec3_t a, float b)
 {
-    return a * (1.0 / b);
+    return a * (1.0f / b);
 }
 
 vec3_t operator*(vec3_t a, vec3_t b)
@@ -58,7 +58,7 @@ vec3_t operator*(vec3_t a, vec3_t b)
     return a;
 }
 
-double dot(vec3_t a, vec3_t b)
+float dot(vec3_t a, vec3_t b)
 {
     return a.x * b.x + a.y * b.y + a.z * b.z;
 }
@@ -72,14 +72,14 @@ vec3_t cross(vec3_t a, vec3_t b)
     );
 }
 
-double magnitude(vec3_t v)
+float magnitude(vec3_t v)
 {
     return sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
 vec3_t normalize(vec3_t v)
 {
-    double len = magnitude(v);
+    float len = magnitude(v);
     if (len == 0.0f)
         return v;
 
@@ -101,18 +101,18 @@ vec3_t reflect(vec3_t i, vec3_t n)
 {
     i = normalize(i);
     n = normalize(n);
-    return normalize(i - 2.0 * n * dot(n, i));
+    return normalize(i - 2.0f * n * dot(n, i));
 }
 
 vec3_t saturate(vec3_t c)
 {
-    c.r = clamp(c.r, 0.0, 1.0);
-    c.g = clamp(c.g, 0.0, 1.0);
-    c.b = clamp(c.b, 0.0, 1.0);
+    c.r = clamp(c.r, 0.0f, 1.0f);
+    c.g = clamp(c.g, 0.0f, 1.0f);
+    c.b = clamp(c.b, 0.0f, 1.0f);
     return c;
 }
 
-double& vec3_t::operator[](int i)
+float& vec3_t::operator[](int i)
 {
     assert(i >= 0 && i < 3 && "Invalid subscript index on vector");
 
@@ -124,7 +124,7 @@ double& vec3_t::operator[](int i)
 
 }
 
-void vec3_t::operator*=(double a)
+void vec3_t::operator*=(float a)
 {
     *this = *this * a;
 }
@@ -147,7 +147,7 @@ void vec3_t::operator-=(vec3_t v)
 
 vec3_t rotate(vec3_t in, vec3_t angles)
 {
-    mat3_t x_rot(0.0), y_rot(0.0), z_rot(0.0);
+    mat3_t x_rot(0.0f), y_rot(0.0f), z_rot(0.0f);
 
     x_rot[0] = vec3_t(1, 0,          0);
     x_rot[1] = vec3_t(0, cos(angles.x), -sin(angles.x));
@@ -165,23 +165,23 @@ vec3_t rotate(vec3_t in, vec3_t angles)
     return in;
 }
 
-static inline double rand_0_1()
+static inline float rand_0_1()
 {
-    double n = (double)std::rand() / RAND_MAX;
+    float n = (float)std::rand() / RAND_MAX;
     return n;
 }
 
 vec3_t get_sphere_random(void)
 {
-    double x, y, z;
-    double m;
+    float x, y, z;
+    float m;
 
     do {
-        x = 2.0 * rand_0_1() - 1.0;
-        y = 2.0 * rand_0_1() - 1.0;
-        z = 2.0 * rand_0_1() - 1.0;
+        x = 2.0f * rand_0_1() - 1.0f;
+        y = 2.0f * rand_0_1() - 1.0f;
+        z = 2.0f * rand_0_1() - 1.0f;
         m = magnitude(vec3_t(x, y, z));
-    } while (m > 1.0);
+    } while (m > 1.0f);
 
     return normalize(vec3_t(x, y, z));
 }
@@ -189,7 +189,7 @@ vec3_t get_sphere_random(void)
 vec3_t get_hemisphere_random(vec3_t dir)
 {
     vec3_t d = get_sphere_random();
-    if (dot(d, dir) < 0.0)
+    if (dot(d, dir) < 0.0f)
         d = -d;
     return d;
 }
